% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isovision.R
\name{isovision}
\alias{isovision}
\title{IsoVision}
\usage{
isovision(
  data,
  whichIsoforms = NULL,
  main = NULL,
  main.cex = 2,
  use.barplot = T,
  barplot.width = 0.1,
  barplot.fill = "white",
  barplot.border = "black",
  line.color = "black",
  intron.reduction = 0.9,
  n.chunks = 1000,
  orf.lwd = 1.5,
  orf.lty = 1,
  norf.lwd = 1,
  norf.lty = 4,
  mar = 0.1,
  terminal.style = "triangle",
  terminal.fill = "white",
  terminal.border = "black",
  default.fill = "white",
  default.border = "black",
  legend.values = setNames(colnames(data)[-(1:8)], colnames(data)[-(1:8)]),
  legend.colors = "Set3",
  filename = "Rplots.pdf",
  width = 11,
  height = 8.5
)
}
\arguments{
\item{data}{Data generated by \code{isoviz_data}.}

\item{whichIsoforms}{A character vector of isoform names or an integer vector for isoform ranks to visualize. If null, plots all isoforms.}

\item{main}{The title.}

\item{main.cex}{The \code{cex} graphical parameter for the title.}

\item{use.barplot}{Whether or not to plot a barplot of total reads next to the track.}

\item{barplot.width}{The relative width of the barplot (0.0 to 1.0).}

\item{barplot.fill}{A vector of colors for the bars. Passed to the R barplot function.}

\item{barplot.border}{A vector of colors for the border of the bars. Passed to the R barplot function.}

\item{line.color}{The color of the track lines as a vector of the same length of the number of isoforms, an RColorBrewer palette or a single value.}

\item{intron.reduction}{The reduction factor for intron widths to ensure long introns are shrunk for plotting.}

\item{n.chunks}{The number of chunks to split the line into for intron scaling. Higher numbers mean more intron reduction.}

\item{orf.lwd}{The border weight for exons in the ORF.}

\item{orf.lty}{The border line type for exons in the ORF.}

\item{norf.lwd}{The border weight for exons not in the ORF.}

\item{norf.lty}{The border line type for exons not in the ORF.}

\item{mar}{The margin.}

\item{terminal.style}{The drawing style for the terminal exons. Options include "arrow" and "triangle". Any other value will draw a normal rectangular exon.}

\item{terminal.fill}{The fill color of the terminal exons.}

\item{terminal.border}{The border color of the terminal exons.}

\item{default.fill}{The fill color of exons without any annotations.}

\item{default.border}{The border color of exons without any annotations.}

\item{legend.values}{A named character with names set to domain annotation names present in \code{data} and values set to the preferred "pretty print" name that will show in the legend.}

\item{legend.colors}{Colors to be applied to domain annotations. Will be applied sequentially in the order that domains are defined in \code{legend.values}. Colors will be recycled if length is less than the number of domains in \code{legend.values}. Can also be an RColorBrewer palette.}

\item{filename}{The name to save this plot as.}

\item{width}{The width, in inches.}

\item{height}{The height, in inches.}
}
\description{
IsoVision
}
