% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoviz_data.R
\name{isoviz_data}
\alias{isoviz_data}
\title{IsoVision Data}
\usage{
isoviz_data(path.bed, path.out, path.rank, path.fasta)
}
\arguments{
\item{path.bed}{The path to the .bed file specifying the feature track.}

\item{path.out}{The path to the .out file from the CDD with conserved domain annotations.}

\item{path.rank}{The path to the SHaMsPipe .tab file specifying isoform ranks.}

\item{path.fasta}{The path to the .fasta file with ORF information in the headers.}
}
\value{
Data ready for plot with IsoVision
}
\description{
IsoVision Data
}
